cache_fetcher: File related methods receive complete file path
Instead of calling the getFilePath multiple times, it receives the
complete cache file path directly. The getFilePath function can be
called once and the result passed as parameter to all other methods.

It avoids potential typos in various calls to file related and makes the
code more consistent: capnpCacheFileExists was often called with
parameters different than those sent to the 'open' call a few lines
after. Now the same variable can be used in both calls.