cache_fetcher: close fd outside reading blocks
fixes #46

When reading capnp data with PackedFdMessageReader, it may happen that
data is accidentally being read after the code is done reading it as it
does buffered I/O. If the file descriptor is closed when that happens,
it causes an exception in the application. (from
https://capnproto.org/cxx.html)

So closing the file descriptor should be done outside the block where
the message reader lives to avoid those exceptions.