Add benchmarks for the CSA algorithm
Fixes #53

This adds a test application to run benchmarks for the CSA algorithm. To
run it, simply run `make check` and it will build this application and
run the benchmarks.

The benchmarks require cache data to be available in the
`tests/connection_scan_algorithm/cache/demo_transition` directory. The
cache data to match the benchmark can be found here:
https://nuage.facil.services/s/ntXfzgfBEFDS7M2 Simply unzip in the
`tests/connection_scan_algorithm/cache` directory. It corresponds to the
STM's fall 2018 18S_S service.

For now, only one query is benchmarked and hard-coded in the file. The
results are outputted in the benchmarkResults.txt file.

TODO:

* Move this benchmark application outside the `make check` target to
instead have a `make bench` target.
* Parameterize cache queries to benchmark multiple queries for the given
cache. Could be a .csv file that comes with a cache archive.
* Document how to do benchmarks.