Add an OpenAPI documentation of the trRouting API
This commit starts a `docs/API.yml` API documentation file using the
OpenAPI 3.0.2 specification. It describe the routing parameters, for now
only those used by the transition query in typescript.

Other parameters are listed in comments in the file, it is to be
discussed whether they are all required or not.

Responses are also detailed in various object types for the main routing
query (with or without alternatives).