Add unit test data for the CSA algorithm
The unit test data was generated with Transition. It's a 2 lines, 2
paths with 5 nodes crossing at their middle, with 2 trips each,
one of which allows for transfer at the mid point. Another line with 1
path has only 2 stops and will allow to test walking transfers from the
midpoint.

The data was exported to gtfs, provided in the code, so it can be imported
in Transition and modified. The times are rounded for easier readability.
The objects are manually created in the unit tests.

This data covers a wide variety of test cases. The setUp* test methods
can be overridden by specific test cases to add
lines/trips/services/etc.