csa: Move node and schedules internal data initialization code
Internal data for nodes and schedules were initialized in the methods
that fetched the nodes and schedules respectively, but this code should
be executed whenever nodes and schedules data are modified, not just
when fetched from cache.

This code is moved to a single initializeCalculationData method in
initializations.cpp file and make public for unit tests to call.