csa: Move the connection setting and sorting to its own method
forward and reverse connection objects are private in calculator, and
need to be of specific format/sorting, etc for the algorithm to work
properly. It is thus not the cache_fetcher's role to prepare this data.

The sorting and setting code is moved from trips and connections cache
fetcher to a Calculator's setConnections method, which resets the
forward and reverse connections data and prepare them for calculation.