osrm_fetcher: Pass the max travel time and walking speed as parameter
Fixes #97

The Parameters object is the server parameters' main object, that
contains the connection information. The speed and max travel time will
soon be extracted out of this class. Also, depending on the calling
context, the max travel time may not be the access time, but the egress
or any other time. This makes these methods more flexible.

Unit test for max egress time is brought back as it now works.