CSA: Let the data preparation also validate the data
Add a DataStatus enum to the calculator with values specifying what is
the status of the data, whether there are missing data, read errors or
if it is ready.

So it is not necessary to validate the agencies/lines/etc count at each
query, the prepare method will return a status code that can then be
used to prepare a response for the query.