Turn the CSA benchmark into a parameterized test
This allows to specify a list of parameters for each calculation to make
and run benchmark for each case.

For each test case, the data is calculated with or without alternatives
and using the time as arrival or departure time.

The results are written to 2 csv files:

* `benchmarkResults.csv` contains a line for each test case and the
results for each combination of alternatives or not and arrival or
departure time (so 4 results).
* `benchmarkResultsDetailed.csv` contains a line per combination with
the individual results for each calculation.