Update googletest's version to 1.11
Use the tagged version `release-1.11.0` instead of the arbitrary commit
that happened to be the head the day we first introduced googletest.

To update, simply `cd googletest`, then use `git fetch origin` and `git
checkout <tag to use>`. Then `cd ..` back to the trRouting directory and
commit the resulting change.