csa: Initialize all fields in the calculator constructor
It is a rule to initialize all fields of a class in a constructor,
otherwise they may be assigned random values that happen to be in memory
when constructing the object.

The main code path was not affected since there's only one instance of
the Calculator constructed at the beginning of the program. But in unit
tests, it is causing bizarre failures of unrelated tests