Document a new endpoint for origin/destination routing queries
This commit is the start of v2 of the trRouting API, where each query
will give access to a single functionality and parameters are
well-defined and for each query, all of the parameters can be set.

This endpoint returns a single transit routing calculation for the
origin/destination in parameter. It is a subset of the route/v1/transit
query, return types are updated to return only the minimal information.

Note that this endpoint reverses the coordinates order of the
parameters. Origin and destination are now lon/lat, instead of lat/lon,
as mentioned in #100.