Change return type of calculation to return an object
The alternatives_routing and calculate methods now return a result
object of the RoutingResultNew class. The alternatives calculation
method returns a concrete object, but since the calculation method
covers both the allNodes and single calculation use case, it is not
possible to determine the concrete type, so a unique_ptr is returned in
this case. As this method is revisited later on, they can be split and
also return concrete types.

The result is then converted to v1 API result in the query. Visitor
classes are used when some result specific data is required, for example,
the lines for alternatives results.

od_trip_routing also use the RoutingResultNew class, but still return a
string, as previously, until its API is better documented.