csa: Pass the query parameters as parameter to the calculation
Instead of updating fields of the calculator object that are specific
for a single query, those fields are part of a RouteParameter object,
which is passed to the calculation method.

This is a first step to make calculations thread-safe, with [almost]
immutable objects (for now the alternative query still needs to update
the exceptLinesIdx vector of the parameter object). There are a few
TODOs on those fields, for the moment we have calculation specific
parameters.

It adds unit tests, with the same query as the original v1 unit tests,
but constructing the routeParameters object directly.