Use forward declaration in .h instead of includes in calculator.hpp
It is a good practice to do so, it avoids including unnecessary headers
in compilation units, it should also decrease compilation times,
especially when make small changes.

Headers are included only in the source files that requires them. This
commit does not significantly reduce compilation time, but the size of
the libcsa.a library has decreased from 107M to 92M.

Also remove a few unnecessary json.hpp includes.