Add reason to "no routing found" in v1 API.
Take the reason enum out of the exception to more easily re-use it.

Add the `reason` field to the no_routing_found api return value, with
reasons corresponding to the ones available for now: no access at origin
and no access at destination.

Update the ResultToV1Response#noRoutingFoundResponse to take reason as
parameter as well.