Remove debugDisplay from the Parameter object
We now exclusively control the debug display with the spdlog loglevel, so no need to propagate
this option everywhere.

In addition to the display, this variable was used to activate some of the benchmarking infrastructure
in a few places. For the moment, the benchmarking will always be initialized until we find a more
dedicated infra for this data. Should not have a significant performance impact