Move OSRM related parameters out of the Parameters class
The OSRM ports and hosts, and also the birdDistanceAccessibilityEnabled field was only used in
the OsrmFetcher static function. They were only initialized at startup.
We move them inside the OsrmFetcher class as static members, which match their usage pattern.

Eventually, it would be better to have an actual OsrmFetcher (or other routing data fetcher) objects,
but that's for a bigger future refactoring