Move cacheDirectoryPath parameter inside the CacheFetcher class
Instead of passing around this information in the Paramaters object, we make it
part of the CacheFetcher object directly. For this, we need to remove the static
qualifier on getFilePath.

Lot of modifications, but the vast majority is to remove the params parameter in
many function call that we don't need anymore