Add const to DataFetcher functions
Several of the parameters passed to the various function were on fact read only
Add const qualifier to highlight and enforce this fact.

Several accessor had to be changed from [] to at(). With that, since at() validate that the data is
there, several tests need to be changed to actually load the dependent data.