Migrate DataSource vector and index map to a single UUID:DataSource map
Added == and != operators to DataSource

Since Person and OdTrip now store a reference to a DataSource, we added constructor for those.

While doing this, we observed that Households and Places are not used anywhere in the calculation.
We have commented out those files, references and initialization functions. Those were marked
with a TODO #167 (task number to restore them when needed/refactoring done)