Migrate Line vector and index map to a single UUID:Line map
There was a few cases where the LineIndex was saved but not used, so those where removed.
This include the ConnectionTuple, which included a LineIdx.

In areas where you had a Trip object, the Line object can be accessed directly instead of
going to the full vector

Combination class had to be changed to be able to accommodate std::reference_wrapper.