Add an API endpoint documentation to get line summary
Documents #177

* Add a API endpoint called `v2/summary`, taking the same parameters as
the `route` endpoint, along with an object to state what to summarize
(only line will be implemented for now). This will return an array of
lines and their agency data, for all lines that were returned by the
route calculation, as well as the number of alternatives they were part
of.
* Move some common responses to the commonResponse.yml file