Implement the `v2/summary` endpoint
Fixes #177

This endpoint does exactly the same calculation as the `v2/route`
endpoint. Only the result converter is changed.

The ResultToV2SummaryResponse converter takes the routing result and
creates the json response with the details for each line used. It also
calculates the number of times a line is encountered in an alternative.