Initialize global parameter object at startup
Initializing those parameters is required for proper calculation as it
still contains a few fields that are not really parameters, but are
there (like accessMode and egressMode for footpath calculations).

This fixes instances planned to run only V2 queries, as this param
object is not re-initialized by those requests, only the
`transit/v1/route` query did it. v2 endpoints all returned no routing
unless a v1 query was done first.