Migrate Transit data storage outside of the Calculator class
Create a TransitData class. This interact with a data fetcher to get
and store the base transit information. All data access with this are const,
so no chances to change them by the calculation.

Effectively split "constant" data and "temporary" variable. (Might have a few other
variable that would be global and can be "cached"/"pre-computed")

Created a TestDataFetcher in the CSA test directory to setup all the mocked data