Split AllNodesResult codepath for forwardJourneyStep
This create a second function forwardJourneyStepAllNodes for the params.returnAllNodesResult == true case.

After multiple pass for unused variable cleanup in the AllNodes case, we ended up with a really simple
function.

We don't have to create the resultingNodes vector. In the allNodes case, we iterate over all nodes
and in the single result one, we just check the bestEgressNode.

Might be possible to optimize further, so far we have only done the most evident ones