Split AllNodesResult codepath for reverseJourneyStep
We could not optimize the AllNodes version as much as the forwardCase as the journey
is reduced by calling optimizeJourney.

We also add a function optmizeCasesToString to convert the optimization case list
to a string. (And move this code out of reverseJourneyStep()

All the other changes are mostly the same than the previous commit for forwardJourneyStep