v2: Add the `v2/accessibility` API route
This route is for the all_nodes of v1, with only the required
parameters, a single place and the time query parameters are the same as
the other calculation routes.

The response type is defined and the node's code and name are returned
additionally, like the other API responses. Consumers don't necessarily
have access to the data and the ID may not be meaningful.