v2: Refactor the reponse for the `route` endpoint
To differentiate the data coming from the query from the ones computed
by trRouting, the response has a `query` and a `result` field.

The alternatives response now return only the calculation result for
each alternative instead of the complete status/query/result. If they
are alternatives, they are successful and the query parameters are the
same for all.