performance: Cache connections by scenario
Fixes #237

This caches the connections for the queried scenario. Only the last
queried scenario is kept to avoid having to manage cache eviction and
memory usage.

At the beginning of the calculation, in the resetFilters step, the
connection cache object is retrieved, and computed if a new scenario
is requested. This object contains only the forward and reverse
connections for this scenario. All others are ignored.

For instances with a lot of services this improves performance. For
example, calculations are 75% faster for an instance that contained the
complete Montreal area data for fall 2018 and 2019.