Calculator: Keep only disabled trips
With the new connection cache, only the connections for the trips
enabled for a scenario are considered in the query. When calculating
alternatives, some trips are disabled at each calculation. Instead of
keeping a map of all trips, only keep the disabled ones for the current
query. This slightly speeds up alternative calculation (slightly)