Change log levels of endpoint entry/exit
Part of #248

The alternatives calculation statements are set to 'debug' instead of
'info'.

The log statement at the beginning and end of the v2/* routes are set to
'info' and parameters are removed, spdlog will associate a timestamp with
the log, so no need to display the application's time.

If a catchall exception occurs, the log level is 'error'.

TODO
Cleanup the 'debug' statements and set a few as 'trace'