csa: Make sure the connection arrival node is a candidate
fixes #250

Lookup the arrival node in the reverse journey steps map to make sure
the arrival node of the connection is a candidate for the journey.
Candidate nodes are added to the map as connections having them as
departure node are found to be possible shortest path candidate.

Otherwise, it throws an exception when calling the `at`, making the
whole query fail.