Add an option to cache all connectionSets from all scenarios
In a situation where you have multiple scenario and where you often switch between
them, having only one connection set in the cache does not offer lot of perf
improvement.

This add a --cacheAllConnectionSets parameter, which when set to true will
cache all connection cache.

This is implemented with a new ScenarioConnectionCacheAll class. (the old one renamed to ScenarioConnectionCacheOne)
The flag define which one the TransitData class will instanciate.

Issue #252