Split osrm_fetcher into 2 classes and a base one
The bird distance was renamed to euclidean.

This allow to pass the desired object to the Calculator constructor and thus allowing
us to select which method we want to use. This replace the global static
variable.

We don't use the cycling and driving mode so we instanciate an OsrmGeoFilter for those. The OSRM
parameters (host/port) are now part of the OsrmGeoFilter object.

AccessMode and EgressMode was always walking, we removed those variables from the Paramters object.