Move the data from the nodesTentativeTime to a mutable inside the node object
Fetching the temporary data from a map in the calculation loop is too intensive.
We this, we move the data inside the node object. Since the basic Node
object is const, we add the variable as mutable. We can use this concept
for more temporary data object.
This concept only work as long as we are single threaded.