Replace the nodesTentativeTime from a map to a vector
We used the node global as a max entry to size the vector.

Fetching the temporary data from a map in the calculation loop is too intensive.
Using a vector with a constant access is more efficient. We just need to be
careful in match the UID of the Nodes