Remove the legacy Parameters class
We moved the remaining parameters to an OdTripLegacyParameters object. Those now only contains
the default values necessary of the od_trips_routing calculation. We don't have a API for it
at the moment, so it's just to keep it compiling.

We removed the calls to setDefaultValues() and removed the params instance in the Calculation object.

We also removed the legacy V1 specific unit tests