Remove the global Calculator object
What's left in the initialization of the Calculator class is lightweight and self contained.
Instead of having a long live Calculator object, we now instanciate an object at the time
of running the appropriate calculation.

We call the initializeCalculationData in the constructor now. (We could merge them later).
We also initialize the algorithmCalculationTime in the constructor, as we assume we will
create the object and run the calculation right away. (And we made it private)