Add an program option to specify the number of threads.
With the latest changes, we can do calculation in parallel in multiple different threads.

The updateCache API call is not thread-safe yet. We need to add locks around the transitData object and the updateCache
call. But in the majority of cases, we do not use this api call, so it's safe to use as is for the moment.