Use external nlohmann-json library instead of copying the files in the project
Distro have packages for the json library that we use. Instead of having to maintain a local
file in the repo, we will rely on external files on the system.

We changed the #include reference to point the the standard files and added a check in the configure
to make sure the file is available.
Also updated the github actions to install the right packages and updated the README to add the package
to the install instructions