Fix build for macos
The homebrew location for boost have changed, it's not in the default search path anymore.

We add the BOOST_CPPFLAGS to our CPPFLAGS
and we update all extra autoconf macro to have the latest update from the autoconf-archive.
The boost packages are updated to search boost in /opt/homebrew
(the update to ax_cxx_compile for C++20 is also included while at it)

Also remove an old comment that should have been removed before in the build workflow