Check for index in path node list when parsing connection
We had an issue when the data got corrupted and the number of entry in the path was not the same
as in the schedule. This was corrupting the memory.
We use the at() instead of [] when accessing the vector in the parsing loop in
getSchedules() and we display a meaningful error message if we
get the exceptiono