Add support to cache accessible nodes via Memcached
Related to issue #271
In the V1 API, Transition had the possibility to send a list of accessible nodes with a request. The
intent for this was to speedup calculation for frequently requested points.

Instead of letting the client send us cache information, we are adding support for TrRouting to
handle the caching directly using Memcached.

When enabled with --useMemcached command-line option, trRouting will compute a key with the point and caracteristic
of the accessible node request (max walking time, speed, direction). If the data was already computed, it will
be fetched from the memcached instance, otherwise it will do the calculation and store the result for later use.

It's implemented as a GeoFilter and receive a geofilter in its constructor, so it can be used with both the
OSRMGeoFilter and the Euclidean one.

In our test, a request to OSRM was taking aroung 50-100ms, and with the caching we got the result in less than 1ms. Not
a huge saving for a single calculation, but still a save. This will be mostly benefical with computations that often check
the same points, like with Transition's simulations. Other type of batch calculation could gain from it if they have
recuring access or egress points.

We currently cache the result for 1h in memcached, to not have stale results and conserve memory over time.

The compilation of this extension is optional. If you don't have libmemcached installed, it won't compile it.
The CLI option is always available and will only display a warning if it was not compiled.

Tested with both OSRM and Euclidean filters, with and without Memcached (including the case when memcached was shutdown)