Use connection pool for memcached connection for thread-safety
The connection object of memcached library is not thread-safe. To insure safe operation, we will
use a connection pool and have one connection for each thread. We use the memcached_pool object
of the libmemcachedutil library. (Part of the same project than libmemcached).

We create a pool of up to the number of configured thread in the application. On a request, we fetch
a connection in the pool, and put in back at the end.

Added the necessary check in configure.ac to detect the libmemcachedutil library. It's part of the
same package usually, so no need to update build instructions.

(Somehow, when we spawn trRouting from Node.JS we need to do it with a shell when we add this library.
Did not find why exactly yet.)