Use Guard pattern to access memcached connection pool.
By using the guard pattern, we tie the lifetime of the connection access and release from the pool
to its usage scope. That way, if an exception arise and we return early from the function, the
connection will be returned to the pool in any cases.

We had long running trRouting process which could not fetch any connection from the pool. The connection were probably leaked due to
unknown errors.