route: calculate best departure time first in reverse calculation
fixes #298

The `calculateSingleReverse` function assumes that the best arrival time
was found and egress footpaths were set to this value before doing
calculation. That is not the case with reverse calculation, where the
arrival time is arbitrarily after the best possible arrival time and the
algorithm was quite creative on the trip end. To avoid this, we first do
a reverse calculation to get the best departure time, then do a normal
forward calculation with that time. That arrival time will be
automatically adjusted.

Add a unit test to reproduce the issue