limit time window at end of trip when time is arrival
fixes #317

Deprecate the `max_first_waiting_time` parameter and replace with
`max_inner_time_of_trip_buffer`, which conveys more the intent of a
maximum buffer time at either start or end of trip. If the time_of_trip
is departure, this value is the maximum waiting time at the first stop
before boarding. If the time_of_trip is arrival, then it is the maximum
time between the actual arrival and the requested time.

Rename all function and variables to the new `maxInnerTimeOfTripbuffer`
name.

Support the inner buffer in the reverse calculation.

Add unit tests that previously failed, but passed with this fix.