route: calculate route with arrival time symetrically with departure
fixes #298

Fix a time instead of distance in the `forwardJourneyStep` method.

Use a similar approach with arrival time calculations as the departure
time one: first get the best access time by a reverse calculation, then
find the best arrival time for this departure time, thus optimizing the
arrival of the route. Then retrieve the journey steps required for the
route.