osrmgeofilter: properly handle `null` distances/durations
fixes #327

When a node is not routable by osrm, the distances and/or durations
returned by the table query can be `null`. When that is the case, we ignore
this node. The `osrm` profiles have been investigated, some parameters of
the profile may make a node effectively inaccessible, either because it
is too far from the network, or the possible routes are not accessible,
so the returned `null` value are valid and represent un-routable points.